refactor(serve): extract embeddable Run entrypoint from the serve command
Move the server implementation out of the ServeCmd.Run CLI handler into a new
pkg/serve package exposing Run(ctx, cfg, ...Option). The CLI command becomes a
thin wrapper that loads configuration from the environment and calls serve.Run,
so the same server can be started by another process that builds its own
ServerConfig instead of reading a YAML file.

No behavior change: the CLI path resolves the storage DSN, ensures the schema,
starts the operator, gRPC server, telemetry, and HTTP server exactly as before.
The Option pattern (WithLogger, WithBuildInfo) is the seam for embedders to
inject their own logger and build identifiers.
