+++ pkg/serve/serve_grpc_test.go
@@ -1,4 +1,4 @@
-package commands
+package serve
 
 import (
 	"log/slog"
